+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/service/DiscordService.kt
@@ -7,15 +7,17 @@ import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.editor.ex.EditorEventMulticasterEx
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
-import de.jcm.discordgamesdk.Core
-import de.jcm.discordgamesdk.CreateParams
-import de.jcm.discordgamesdk.activity.Activity
 import io.github.pandier.intellijdiscordrp.DiscordRichPresencePlugin
 import io.github.pandier.intellijdiscordrp.activity.ActivityContext
 import io.github.pandier.intellijdiscordrp.activity.currentActivityApplicationType
 import io.github.pandier.intellijdiscordrp.listener.RichPresenceFocusChangeListener
 import io.github.pandier.intellijdiscordrp.settings.discordSettingsComponent
+import io.github.pandier.intellijdiscordrp.util.KPresenceLoggerAdapter
 import io.github.pandier.intellijdiscordrp.util.MergingRunner
+import io.github.vyfor.kpresence.RichClient
+import io.github.vyfor.kpresence.exception.NotConnectedException
+import io.github.vyfor.kpresence.exception.PipeNotFoundException
+import io.github.vyfor.kpresence.rpc.Activity
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
@@ -23,7 +25,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
-private fun connect(): Core {
+private fun connect(): RichClient {
     val settings = discordSettingsComponent.state
     val applicationId = if (settings.customApplicationIdEnabled) {
         settings.customApplicationId.toULong().toLong()
@@ -31,12 +33,10 @@ private fun connect(): Core {
         currentActivityApplicationType.discordApplicationId
     }
 
-    val internal = Core(CreateParams().apply {
-        clientID = applicationId
-        flags = CreateParams.getDefaultFlags()
-    })
-
-    return internal
+    return RichClient(applicationId).apply {
+        logger = KPresenceLoggerAdapter
+        connect()
+    }
 }
 
 /**
@@ -59,7 +59,7 @@ class DiscordService(
     /**
      * A connection with the Discord client.
      */
-    private var connection: Core? = null
+    private var connection: RichClient? = null
 
     /**
      * The latest [ActivityContext] that was changed.
@@ -69,7 +69,7 @@ class DiscordService(
     /**
      * A [MergingRunner] for the [reconnect] function.
      */
-    private val reconnectRunner = MergingRunner<Unit>(scope, Dispatchers.IO)
+    private val reconnectRunner = MergingRunner<Boolean>()
 
     init {
         // Register focus change listener
@@ -78,43 +78,51 @@ class DiscordService(
         eventMulticasterEx?.addFocusChangeListener(RichPresenceFocusChangeListener, this)
 
         // Connect to Discord client
-        scope.launch(Dispatchers.IO) {
-            try {
-                reconnect().await()
-            } catch (ex: Exception) {
-                DiscordRichPresencePlugin.logger.info("Failed to connect to Discord client", ex)
-            }
-        }
+        reconnectBackground()
     }
 
     /**
-     * Starts the reconnection process or merges into an existing one.
+     * Executes the reconnection process in the current coroutine or merges into an existing process.
      * The reconnection process consists of closing the connection and starting a new one.
      * It's executed using [MergingRunner].
+     *
+     * When [silent] is true, only logs related to state change will be logged.
      */
-    suspend fun reconnect(): Deferred<Unit> = reconnectRunner.run {
-        mutex.withLock {
-            connection?.close()
-            connection = null
-        }
-        val newConnection = connect()
-        mutex.withLock {
-            connection = newConnection
-        }
-        update()
+    suspend fun reconnect(silent: Boolean = false): Deferred<Boolean> = reconnectRunner.run {
+        try {
+            mutex.withLock {
+                connection?.shutdown()
+                connection = null
+            }
 
-        DiscordRichPresencePlugin.logger.info("Connected to Discord client")
+            val newConnection = connect()
+            mutex.withLock {
+                connection = newConnection
+                sendActivityInternal(activityContext?.createActivity())
+            }
+
+            DiscordRichPresencePlugin.logger.info("Connected to Discord client")
+            true
+        } catch (ex: PipeNotFoundException) {
+            if (!silent)
+                DiscordRichPresencePlugin.logger.info("Could not find any Discord client instance")
+            false
+        } catch (ex: Exception) {
+            if (!silent)
+                DiscordRichPresencePlugin.logger.error("Failed to connect", ex)
+            throw ex
+        }
     }
 
     /**
-     * Reconnects on the background silently in a coroutine with [Dispatchers.IO] context.
+     * Executes the reconnection process on the background or merges into an existing one.
      *
      * @see reconnect
      */
-    fun reconnectBackground() {
+    fun reconnectBackground(silent: Boolean = false) {
         scope.launch(Dispatchers.IO) {
             @Suppress("DeferredResultUnused")
-            reconnect()
+            reconnect(silent)
         }
     }
 
@@ -214,8 +222,7 @@ class DiscordService(
         }
 
         if (!success && discordSettingsComponent.state.reconnectOnUpdate) {
-            @Suppress("DeferredResultUnused")
-            reconnect()
+            reconnectBackground(true)
         }
     }
 
@@ -241,16 +248,18 @@ class DiscordService(
 
     private fun sendActivityInternal(activity: Activity?): Boolean {
         return try {
-            connection?.activityManager()?.updateActivity(activity) != null
-        } catch (ex: RuntimeException) {
+            connection?.update(activity) != null
+        } catch (ex: NotConnectedException) {
+            connection = null
+            false
+        } catch (ex: Exception) {
             DiscordRichPresencePlugin.logger.info("An error ocurred while sending activity", ex)
-            connection?.close()
             connection = null
             false
         }
     }
 
     override fun dispose() {
-        connection?.close()
+        connection?.shutdown()
     }
 }